fix: abstract TradeFooter
