+++ packages/maskbook/src/plugins/Trader/UI/trending/TrendingViewDeck.tsx
@@ -2,12 +2,10 @@ import { useCallback } from 'react'
 import {
     Avatar,
     Button,
-    CardActions,
     CardContent,
     CardHeader,
     createStyles,
     IconButton,
-    Link,
     makeStyles,
     Paper,
     Typography,
@@ -15,14 +13,12 @@ import {
 import MonetizationOnOutlinedIcon from '@material-ui/icons/MonetizationOnOutlined'
 import ArrowDropDownIcon from '@material-ui/icons/ArrowDropDown'
 import stringify from 'json-stable-stringify'
-import { findIndex, first, last } from 'lodash-es'
+import { first, last } from 'lodash-es'
 import { Coin, Currency, DataProvider, Stat, TradeProvider, Trending } from '../../types'
-import { resolveDataProviderName, resolveTradeProviderName } from '../../pipes'
 import { formatCurrency } from '../../../Wallet/formatter'
 import { PriceChanged } from './PriceChanged'
 import { Linking } from './Linking'
 import { useI18N } from '../../../../utils/i18n-next-ui'
-import { MaskbookTextIcon } from '../../../../resources/MaskbookIcon'
 import { TrendingCard, TrendingCardProps } from './TrendingCard'
 import { useRemoteControlledDialog } from '../../../../utils/hooks/useRemoteControlledDialog'
 import { PluginTransakMessages } from '../../../Transak/messages'
@@ -30,10 +26,8 @@ import { useAccount } from '../../../../web3/hooks/useAccount'
 import { Flags } from '../../../../utils/flags'
 import { TokenIcon } from '../../../../extension/options-page/DashboardComponents/TokenIcon'
 import { useStylesExtends } from '../../../../components/custom-ui-helper'
-import { FootnoteMenu, FootnoteMenuOption } from '../trader/FootnoteMenu'
-import { getEnumAsArray } from '../../../../utils/enum'
-import { TradeProviderIcon } from '../trader/TradeProviderIcon'
-import { DataProviderIcon } from '../trader/DataProviderIcon'
+import type { FootnoteMenuOption } from '../trader/FootnoteMenu'
+import { TradeFooter } from '../trader/TradeFooter'
 import {
     currentTrendingDataProviderSettings,
     currentTradeProviderSettings,
@@ -65,9 +59,6 @@ const useStyles = makeStyles((theme) => {
             position: 'relative',
         },
         body: {},
-        footer: {
-            justifyContent: 'space-between',
-        },
         headline: {
             display: 'flex',
             alignItems: 'center',
@@ -102,27 +93,6 @@ const useStyles = makeStyles((theme) => {
             color: theme.palette.text.secondary,
             fontWeight: 300,
             marginRight: theme.spacing(1),
-        },
-        footnote: {
-            color: theme.palette.text.secondary,
-            fontSize: 10,
-            marginRight: theme.spacing(0.5),
-        },
-        footLink: {
-            cursor: 'pointer',
-            marginRight: theme.spacing(0.5),
-            '&:last-child': {
-                marginRight: 0,
-            },
-        },
-        footMenu: {
-            color: theme.palette.text.secondary,
-            fontSize: 10,
-            display: 'flex',
-            alignItems: 'center',
-        },
-        footName: {
-            marginLeft: theme.spacing(0.5),
         },
         avatar: {
             backgroundColor: theme.palette.common.white,
@@ -208,8 +178,6 @@ export function TrendingViewDeck(props: TrendingViewDeckProps) {
     )
     //#endregion
 
-    const dataProviderOptions = getEnumAsArray(DataProvider).filter((x) => dataProviders.includes(x.value))
-    const tradeProviderOptions = getEnumAsArray(TradeProvider).filter((x) => tradeProviders.includes(x.value))
     const { approvedTokens, onApprove } = useApprovedTokens(trending.coin.eth_address)
     return (
         <TrendingCard {...TrendingCardProps}>
@@ -301,58 +269,16 @@ export function TrendingViewDeck(props: TrendingViewDeckProps) {
                 </Paper>
             </CardContent>
 
-            <CardActions className={classes.footer}>
-                <Typography className={classes.footnote} variant="subtitle2">
-                    <span>Powered by </span>
-                    <Link
-                        className={classes.footLink}
-                        color="textSecondary"
-                        target="_blank"
-                        rel="noopener noreferrer"
-                        title="Mask"
-                        href="https://mask.io">
-                        <MaskbookTextIcon classes={{ root: classes.maskbook }} viewBox="0 0 80 20" />
-                    </Link>
-                </Typography>
-                {showDataProviderIcon ? (
-                    <div className={classes.footMenu}>
-                        <Typography className={classes.footnote}>Data Source</Typography>
-                        <FootnoteMenu
-                            options={dataProviderOptions.map((x) => ({
-                                name: (
-                                    <>
-                                        <DataProviderIcon provider={x.value} />
-                                        <span className={classes.footName}>{resolveDataProviderName(x.value)}</span>
-                                    </>
-                                ),
-                                value: x.value,
-                            }))}
-                            selectedIndex={findIndex(dataProviderOptions, (x) => x.value === dataProvider)}
-                            onChange={onDataProviderChange}
-                        />
-                        <ArrowDropDownIcon />
-                    </div>
-                ) : null}
-                {showTradeProviderIcon ? (
-                    <div className={classes.footMenu}>
-                        <Typography className={classes.footnote}>Supported by</Typography>
-                        <FootnoteMenu
-                            options={tradeProviderOptions.map((x) => ({
-                                name: (
-                                    <>
-                                        <TradeProviderIcon provider={x.value} />
-                                        <span className={classes.footName}>{resolveTradeProviderName(x.value)}</span>
-                                    </>
-                                ),
-                                value: x.value,
-                            }))}
-                            selectedIndex={findIndex(getEnumAsArray(TradeProvider), (x) => x.value === tradeProvider)}
-                            onChange={onTradeProviderChange}>
-                            <ArrowDropDownIcon />
-                        </FootnoteMenu>
-                    </div>
-                ) : null}
-            </CardActions>
+            <TradeFooter
+                showDataProviderIcon={showDataProviderIcon}
+                showTradeProviderIcon={showTradeProviderIcon}
+                dataProvider={dataProvider}
+                tradeProvider={tradeProvider}
+                dataProviders={dataProviders}
+                tradeProviders={tradeProviders}
+                onDataProviderChange={onDataProviderChange}
+                onTradeProviderChange={onTradeProviderChange}
+            />
         </TrendingCard>
     )
 }
